[FrameworkBundle][Serializer] Deprecate annotations
