+++ src/Symfony/Bridge/Twig/composer.json
@@ -43,7 +43,7 @@
         "symfony/security-core": "^5.4|^6.0|^7.0",
         "symfony/security-csrf": "^5.4|^6.0|^7.0",
         "symfony/security-http": "^5.4|^6.0|^7.0",
-        "symfony/serializer": "^6.2|^7.0",
+        "symfony/serializer": "^6.4|^7.0",
         "symfony/stopwatch": "^5.4|^6.0|^7.0",
         "symfony/console": "^5.4|^6.0|^7.0",
         "symfony/expression-language": "^5.4|^6.0|^7.0",
@@ -61,7 +61,7 @@
         "symfony/http-foundation": "<5.4",
         "symfony/http-kernel": "<6.4",
         "symfony/mime": "<6.2",
-        "symfony/serializer": "<6.2",
+        "symfony/serializer": "<6.4",
         "symfony/translation": "<5.4",
         "symfony/workflow": "<5.4"
     },
