+++ src/Symfony/Component/HttpKernel/composer.json
@@ -66,7 +66,7 @@
         "symfony/translation": "<5.4",
         "symfony/translation-contracts": "<2.5",
         "symfony/twig-bridge": "<5.4",
-        "symfony/validator": "<5.4",
+        "symfony/validator": "<6.4",
         "symfony/var-dumper": "<6.3",
         "twig/twig": "<2.13"
     },
