+++ src/Symfony/Component/PropertyInfo/composer.json
@@ -27,7 +27,7 @@
         "symfony/string": "^5.4|^6.0|^7.0"
     },
     "require-dev": {
-        "symfony/serializer": "^5.4|^6.0|^7.0",
+        "symfony/serializer": "^6.4|^7.0",
         "symfony/cache": "^5.4|^6.0|^7.0",
         "symfony/dependency-injection": "^5.4|^6.0|^7.0",
         "phpdocumentor/reflection-docblock": "^5.2",
@@ -37,7 +37,7 @@
         "phpdocumentor/reflection-docblock": "<5.2",
         "phpdocumentor/type-resolver": "<1.5.1",
         "symfony/dependency-injection": "<5.4",
-        "symfony/serializer": "<5.4"
+        "symfony/serializer": "<6.4"
     },
     "autoload": {
         "psr-4": { "Symfony\\Component\\PropertyInfo\\": "" },
