+++ src/Symfony/Component/Serializer/CHANGELOG.md
@@ -6,11 +6,11 @@ CHANGELOG
 
  * Allow `Context` attribute to target classes
  * Deprecate Doctrine annotations support in favor of native attributes
- * Deprecate passing an annotation reader to the constructor of `AnnotationLoader`
  * Allow the `Groups` attribute/annotation on classes
  * JsonDecode: Add `json_decode_detailed_errors` option
  * Make `ProblemNormalizer` give details about Messenger's `ValidationFailedException`
  * Add `XmlEncoder::CDATA_WRAPPING` context option
+ * Deprecate `AnnotationLoader`, use `AttributeLoader` instead
 
 6.3
 ---
