+++ src/Symfony/Component/Serializer/Tests/Mapping/Factory/ClassMetadataFactoryTest.php
@@ -14,7 +14,7 @@ namespace Symfony\Component\Serializer\Tests\Mapping\Factory;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Serializer\Mapping\Factory\ClassMetadataFactory;
 use Symfony\Component\Serializer\Mapping\Factory\ClassMetadataFactoryInterface;
-use Symfony\Component\Serializer\Mapping\Loader\AnnotationLoader;
+use Symfony\Component\Serializer\Mapping\Loader\AttributeLoader;
 use Symfony\Component\Serializer\Mapping\Loader\LoaderChain;
 use Symfony\Component\Serializer\Tests\Fixtures\Attributes\GroupDummy;
 use Symfony\Component\Serializer\Tests\Fixtures\Attributes\GroupDummyInterface;
@@ -34,7 +34,7 @@ class ClassMetadataFactoryTest extends TestCase
 
     public function testGetMetadataFor()
     {
-        $factory = new ClassMetadataFactory(new AnnotationLoader());
+        $factory = new ClassMetadataFactory(new AttributeLoader());
         $classMetadata = $factory->getMetadataFor(GroupDummy::class);
 
         $this->assertEquals(TestClassMetadataFactory::createClassMetadata('Symfony\Component\Serializer\Tests\Fixtures\Attributes', true, true), $classMetadata);
@@ -42,7 +42,7 @@ class ClassMetadataFactoryTest extends TestCase
 
     public function testHasMetadataFor()
     {
-        $factory = new ClassMetadataFactory(new AnnotationLoader());
+        $factory = new ClassMetadataFactory(new AttributeLoader());
         $this->assertTrue($factory->hasMetadataFor(GroupDummy::class));
         $this->assertTrue($factory->hasMetadataFor(GroupDummyParent::class));
         $this->assertTrue($factory->hasMetadataFor(GroupDummyInterface::class));
