+++ src/Symfony/Component/Serializer/Tests/NameConverter/MetadataAwareNameConverterTest.php
@@ -17,7 +17,7 @@ use Symfony\Component\Serializer\Annotation\SerializedPath;
 use Symfony\Component\Serializer\Exception\LogicException;
 use Symfony\Component\Serializer\Mapping\Factory\ClassMetadataFactory;
 use Symfony\Component\Serializer\Mapping\Factory\ClassMetadataFactoryInterface;
-use Symfony\Component\Serializer\Mapping\Loader\AnnotationLoader;
+use Symfony\Component\Serializer\Mapping\Loader\AttributeLoader;
 use Symfony\Component\Serializer\NameConverter\MetadataAwareNameConverter;
 use Symfony\Component\Serializer\NameConverter\NameConverterInterface;
 use Symfony\Component\Serializer\Tests\Fixtures\Attributes\SerializedNameDummy;
@@ -40,7 +40,7 @@ final class MetadataAwareNameConverterTest extends TestCase
      */
     public function testNormalize(string|int $propertyName, string|int $expected)
     {
-        $classMetadataFactory = new ClassMetadataFactory(new AnnotationLoader());
+        $classMetadataFactory = new ClassMetadataFactory(new AttributeLoader());
 
         $nameConverter = new MetadataAwareNameConverter($classMetadataFactory);
 
@@ -52,7 +52,7 @@ final class MetadataAwareNameConverterTest extends TestCase
      */
     public function testNormalizeWithFallback(string|int $propertyName, string|int $expected)
     {
-        $classMetadataFactory = new ClassMetadataFactory(new AnnotationLoader());
+        $classMetadataFactory = new ClassMetadataFactory(new AttributeLoader());
 
         $fallback = $this->createMock(NameConverterInterface::class);
         $fallback
@@ -70,7 +70,7 @@ final class MetadataAwareNameConverterTest extends TestCase
      */
     public function testDenormalize(string|int $expected, string|int $propertyName)
     {
-        $classMetadataFactory = new ClassMetadataFactory(new AnnotationLoader());
+        $classMetadataFactory = new ClassMetadataFactory(new AttributeLoader());
 
         $nameConverter = new MetadataAwareNameConverter($classMetadataFactory);
 
@@ -82,7 +82,7 @@ final class MetadataAwareNameConverterTest extends TestCase
      */
     public function testDenormalizeWithFallback(string|int $expected, string|int $propertyName)
     {
-        $classMetadataFactory = new ClassMetadataFactory(new AnnotationLoader());
+        $classMetadataFactory = new ClassMetadataFactory(new AttributeLoader());
 
         $fallback = $this->createMock(NameConverterInterface::class);
         $fallback
@@ -120,7 +120,7 @@ final class MetadataAwareNameConverterTest extends TestCase
      */
     public function testNormalizeWithGroups(string $propertyName, string $expected, array $context = [])
     {
-        $classMetadataFactory = new ClassMetadataFactory(new AnnotationLoader());
+        $classMetadataFactory = new ClassMetadataFactory(new AttributeLoader());
 
         $nameConverter = new MetadataAwareNameConverter($classMetadataFactory);
 
@@ -132,7 +132,7 @@ final class MetadataAwareNameConverterTest extends TestCase
      */
     public function testDenormalizeWithGroups(string $expected, string $propertyName, array $context = [])
     {
-        $classMetadataFactory = new ClassMetadataFactory(new AnnotationLoader());
+        $classMetadataFactory = new ClassMetadataFactory(new AttributeLoader());
 
         $nameConverter = new MetadataAwareNameConverter($classMetadataFactory);
 
@@ -154,7 +154,7 @@ final class MetadataAwareNameConverterTest extends TestCase
 
     public function testDenormalizeWithCacheContext()
     {
-        $classMetadataFactory = new ClassMetadataFactory(new AnnotationLoader());
+        $classMetadataFactory = new ClassMetadataFactory(new AttributeLoader());
 
         $nameConverter = new MetadataAwareNameConverter($classMetadataFactory);
 
@@ -165,7 +165,7 @@ final class MetadataAwareNameConverterTest extends TestCase
 
     public function testDenormalizeWithNestedPathAndName()
     {
-        $classMetadataFactory = new ClassMetadataFactory(new AnnotationLoader());
+        $classMetadataFactory = new ClassMetadataFactory(new AttributeLoader());
         $nameConverter = new MetadataAwareNameConverter($classMetadataFactory);
         $this->expectException(LogicException::class);
         $this->expectExceptionMessage('Found SerializedName and SerializedPath annotations on property "foo" of class "Symfony\Component\Serializer\Tests\NameConverter\NestedPathAndName".');
@@ -174,7 +174,7 @@ final class MetadataAwareNameConverterTest extends TestCase
 
     public function testNormalizeWithNestedPathAndName()
     {
-        $classMetadataFactory = new ClassMetadataFactory(new AnnotationLoader());
+        $classMetadataFactory = new ClassMetadataFactory(new AttributeLoader());
         $nameConverter = new MetadataAwareNameConverter($classMetadataFactory);
         $this->expectException(LogicException::class);
         $this->expectExceptionMessage('Found SerializedName and SerializedPath annotations on property "foo" of class "Symfony\Component\Serializer\Tests\NameConverter\NestedPathAndName".');
